Require composer install, update SDK
Fixes #1

The SDK has been updated a few times since this was written, so we should keep up with the changes. Also, the `vendor/` library should not have been committed to the repository. This updates it to lock in to the `v3.4*` branch of the SDK, add the composer install step and updates `index.php` so it uses the correct class names. :sparkles: